@@ -30,72 +30,72 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.0",
-    "@angular/common": "~12.0.0",
-    "@angular/compiler": "~12.0.0",
-    "@angular/core": "~12.0.0",
-    "@angular/forms": "~12.0.0",
-    "@angular/platform-browser": "~12.0.0",
-    "@angular/platform-browser-dynamic": "~12.0.0",
-    "@angular/router": "~12.0.0",
+    "@angular/animations": "~12.0.1",
+    "@angular/common": "~12.0.1",
+    "@angular/compiler": "~12.0.1",
+    "@angular/core": "~12.0.1",
+    "@angular/forms": "~12.0.1",
+    "@angular/platform-browser": "~12.0.1",
+    "@angular/platform-browser-dynamic": "~12.0.1",
+    "@angular/router": "~12.0.1",
     "@ngneat/lib": "^2.0.0",
     "@ngrx/effects": "^12.0.0",
     "@ngrx/router-store": "^12.0.0",
     "@ngrx/store": "^12.0.0",
-    "@reduxjs/toolkit": "^1.6.0-alpha.2",
-    "rxjs": "~7.0.1",
+    "@reduxjs/toolkit": "^1.6.0-beta.0",
+    "rxjs": "~7.1.0",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^11.2.0",
-    "@angular-devkit/build-angular": "~12.0.0",
-    "@angular-eslint/builder": "^4.3.0",
-    "@angular-eslint/eslint-plugin": "^4.3.0",
-    "@angular-eslint/eslint-plugin-template": "^4.3.0",
-    "@angular-eslint/schematics": "^4.3.0",
-    "@angular-eslint/template-parser": "^4.3.0",
-    "@angular/cli": "~12.0.0",
-    "@angular/compiler-cli": "~12.0.0",
+    "@angular-builders/jest": "^12.0.0",
+    "@angular-devkit/build-angular": "~12.0.1",
+    "@angular-eslint/builder": "^12.0.0",
+    "@angular-eslint/eslint-plugin": "^12.0.0",
+    "@angular-eslint/eslint-plugin-template": "^12.0.0",
+    "@angular-eslint/schematics": "^12.0.0",
+    "@angular-eslint/template-parser": "^12.0.0",
+    "@angular/cli": "~12.0.1",
+    "@angular/compiler-cli": "~12.0.1",
     "@commitlint/cli": "^12.1.4",
     "@commitlint/config-conventional": "^12.1.4",
     "@ngrx/store-devtools": "^12.0.0",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@tailwindcss/forms": "^0.3.2",
-    "@testing-library/angular": "^10.6.0",
+    "@testing-library/angular": "^10.7.1",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/user-event": "^13.1.9",
     "@types/jest": "^26.0.23",
     "@types/node": "^12.20.13",
     "@types/node-fetch": "^2.5.10",
-    "@typescript-eslint/eslint-plugin": "^4.23.0",
-    "@typescript-eslint/parser": "^4.23.0",
+    "@typescript-eslint/eslint-plugin": "^4.24.0",
+    "@typescript-eslint/parser": "^4.24.0",
     "all-contributors-cli": "^6.20.0",
     "angular-cli-ghpages": "^1.0.0-rc.1",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.26.0",
+    "eslint": "^7.27.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-import": "2.23.3",
     "eslint-plugin-jest": "^24.3.6",
     "eslint-plugin-jest-dom": "^3.9.0",
-    "eslint-plugin-jsdoc": "34.0.2",
-    "eslint-plugin-ngrx": "^1.21.1",
+    "eslint-plugin-jsdoc": "35.0.0",
+    "eslint-plugin-ngrx": "^1.23.1",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-rxjs": "^3.3.0",
-    "eslint-plugin-testing-library": "^4.3.0",
+    "eslint-plugin-rxjs": "^3.3.1",
+    "eslint-plugin-testing-library": "^4.6.0",
     "git-cz": "^4.7.6",
     "husky": "^6.0.0",
     "jest": "^26.6.3",
     "jest-preset-angular": "^8.4.0",
     "jsonc-parser": "^3.0.0",
     "lint-staged": "^11.0.0",
-    "msw": "^0.28.2",
+    "msw": "^0.29.0",
     "ng-packagr": "^12.0.0",
     "node-fetch": "^2.6.1",
-    "postcss": "^8.2.15",
+    "postcss": "^8.3.0",
     "prettier": "^2.3.0",
     "protractor": "~7.0.0",
     "semantic-release": "^17.4.3",
